Support for physical and logical replication
This patch was implemented by Petros Angelatos and Jeff Davis
to support physical and logical replication in rust-postgres
(see https://github.com/sfackler/rust-postgres/pull/752).

The original PR never made it to the upstream, but we
(Neon) still use it in our own fork of rust-postgres.

The following commits were squashed together:

* Image configuration updates.

* Make simple_query::encode() pub(crate).

* decoding logic for replication protocol

* Connection string config for replication.

* add copy_both_simple method

* helper ReplicationStream type for replication protocol

This can be optionally used with a CopyBoth stream to decode the
replication protocol

* decoding logic for logical replication protocol

* helper LogicalReplicationStream type to decode logical replication

* add postgres replication integration test

* add simple query versions of copy operations

* replication: use SystemTime for timestamps at API boundary

Co-authored-by: Petros Angelatos <petrosagg@gmail.com>
Co-authored-by: Jeff Davis <jeff@j-davis.com>
Co-authored-by: Dmitry Ivanov <ivadmi5@gmail.com>